+++ application/src/main/java/com/foodielog/application/reply/controller/ReplyController.java
@@ -1,10 +1,9 @@
 package com.foodielog.application.reply.controller;
 
-import com.foodielog.application.reply.dto.ReplyCreateParam;
 import com.foodielog.application.reply.dto.ReplyCreateReq;
 import com.foodielog.application.reply.dto.ReportReplyReq;
-import com.foodielog.application.reply.service.dto.ReplyCreateResp;
 import com.foodielog.application.reply.service.ReplyService;
+import com.foodielog.application.reply.service.dto.ReplyCreateResp;
 import com.foodielog.server._core.security.auth.PrincipalDetails;
 import com.foodielog.server._core.util.ApiUtils;
 import com.foodielog.server.user.entity.User;
